requirements.txt
riscv-reg-block
pytest-cov
streamlit
plotly
networkx
pylint
streamlit_autorefresh